Initial commit
When imported as is into IDEA it fails - IDEA complains that it cannot find
the method `buildCache` on `DefaultSettings`. When removing the
`buildCache` call from `buildSrc/settings.gradle` then the import works.
There is also a `buildCache` block in `settings.gradle`, but this one seems
to work even though the one in `buildSrc` does not.
The wrapper version is recent enough to support the buildCache block.